chore: move Navigation impls under spi/
Fixes #4457
part of #4364
